dsc/requirements.txt
aiohttp
discord.py
beautifulsoup4
aiocache
python-dotenv
актуальные версии